# swagger.yaml
openapi: 3.0.3
info:
  title: Image Classification API
  version: "1.0.0"
servers:
  - description: Local server
    url: http://localhost:80
  - description: Local server (alternative)
    url: http://localhost:5000
  - description: Development server
    url: http://172.20.10.2 
  - description: Remote Azure server
    url: https://smart-trash.azurewebsites.net/
...
paths:
...
  /api/images/predict:
    post:
      tags: [Classification]
      summary: Predict labels for provided images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [images, device_name]
              properties:
                images:
                  type: array
                  description: Base64-encoded images (non-empty)
                  items:
                    type: string
                    format: byte
                device_name:
                  type: string
                save_images:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Predictions per image (ordered by confidence); each inner list is `[label, confidence]`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionScores'
        "400":
          description: Missing images or device name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/images/save:
    post:
      tags: [Images]
      summary: Predict and persist images with predictions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [images, device_name]
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: byte
                device_name:
                  type: string
                save_images:
                  type: boolean
                  description: Ignored (images are always saved here)
      responses:
        "201":
          description: Images persisted with predictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/resources/images/{image_path}:
    get:
      tags: [Images]
      summary: Retrieve a saved image by path
      parameters:
        - in: path
          name: image_path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/device/regeister:
    post:
      tags: [Devices]
      summary: Register a device by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_name]
              properties:
                device_name:
                  type: string
                location:
                  type: string
                  default: Unknown
      responses:
        "201":
          description: Device registered (idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        "400":
          description: Missing device name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/update_device_status:
    post:
      tags: [Devices]
      summary: Update device online status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_name, device_up]
              properties:
                device_name:
                  type: string
                device_up:
                  type: boolean
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "400":
          description: Missing device name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/image/{image_id}:
    get:
      tags: [Images]
      summary: Get image metadata by id
      parameters:
        - in: path
          name: image_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Image record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    put:
      tags: [Images]
      summary: Partially update image metadata
      parameters:
        - in: path
          name: image_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageUpdate'
      responses:
        "200":
          description: Updated image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    delete:
      tags: [Images]
      summary: Delete image
      parameters:
        - in: path
          name: image_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
...
  /api/device/{device_id}:
    get:
      tags: [Devices]
      summary: Get device
      parameters:
        - in: path
          name: device_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    put:
      tags: [Devices]
      summary: Update device (partial)
      parameters:
        - in: path
          name: device_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    delete:
      tags: [Devices]
      summary: Delete device
      parameters:
        - in: path
          name: device_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
...
  /api/prediction/{prediction_id}:
    get:
      tags: [Predictions]
      summary: Get prediction
      parameters:
        - in: path
          name: prediction_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prediction record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prediction'
        "404":
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    put:
      tags: [Predictions]
      summary: Update prediction (partial)
      parameters:
        - in: path
          name: prediction_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionUpdate'
      responses:
        "200":
          description: Updated prediction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prediction'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    delete:
      tags: [Predictions]
      summary: Delete prediction
      parameters:
        - in: path
          name: prediction_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/prediction:
    post:
      tags: [Predictions]
      summary: Create prediction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionCreate'
      responses:
        "201":
          description: Created prediction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prediction'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
...
  /api/fill_level/{fill_level_id}:
    get:
      tags: ["Fill Levels"]
      summary: Get fill level
      parameters:
        - in: path
          name: fill_level_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fill level record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FillLevel'
        "404":
          description: Fill level not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    put:
      tags: ["Fill Levels"]
      summary: Update fill level (partial)
      parameters:
        - in: path
          name: fill_level_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FillLevelUpdate'
      responses:
        "200":
          description: Updated fill level
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FillLevel'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Fill level not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    delete:
      tags: ["Fill Levels"]
      summary: Delete fill level
      parameters:
        - in: path
          name: fill_level_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: Fill level not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/fill_level:
    post:
      tags: ["Fill Levels"]
      summary: Create fill level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FillLevel'
      responses:
        "201":
          description: Created fill level
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FillLevel'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
...
  /api/fill_level/device/{device_id}:
    get:
      tags: ["Fill Levels"]
      summary: Get fill levels by device
      parameters:
        - in: path
          name: device_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fill levels for device
          content:
            application/json:
              schema:
                type: object
                properties:
                  fill_level:
                    type: array
                    items:
                      $ref: '#/components/schemas/Prediction'
        "404":
          description: Device not found or no predictions for device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/user/{user_id}:
    get:
      tags: [Users]
      summary: Get user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    put:
      tags: [Users]
      summary: Update user (partial)
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/user:
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
...
  /api/login:
    post:
      tags: [Users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
  /api/logout:
    post:
      tags: [Users]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiMessage'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiMessage:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties: true
    PredictionScores:
      type: array
      description: List of predictions per image; follows the order of the submitted images
      items:
        $ref: '#/components/schemas/PredictionList'
...
    Device:
      type: object
      properties:
        id:
          type: integer
        device_name:
          type: string
        location:
          type: string
          nullable: true
        device_up:
          type: boolean
          default: false
        last_update:
          type: string
          format: date-time
          nullable: true
      required: [device_name]
    DeviceUpdate:
      type: object
      properties:
        device_name:
          type: string
        location:
          type: string
          nullable: true
        device_up:
          type: boolean
        last_update:
          type: string
          format: date-time
          nullable: true
    Image:
      type: object
      properties:
        id:
          type: integer
        path:
          type: string
        width:
          type: integer
        height:
          type: integer
        creation_date:
          type: string
          format: date-time
      required: [path, width, height, creation_date]
    ImageUpdate:
      type: object
      properties:
        path:
          type: string
        width:
          type: integer
        height:
          type: integer
        creation_date:
          type: string
          format: date-time
    Prediction:
      type: object
      properties:
        id:
          type: integer
        image_id:
          type: integer
        device_id:
          type: integer
        prediction_label:
          $ref: '#/components/schemas/PredictionLabel'
        confidence:
          type: number
          format: float
        real_label:
          $ref: '#/components/schemas/PredictionLabel'
          nullable: true
      required: [image_id, device_id, prediction_label, confidence]
    PredictionCreate:
      type: object
      required: [image_id, device_id, prediction_label, confidence]
      properties:
        image_id:
          type: integer
        device_id:
          type: integer
        prediction_label:
          $ref: '#/components/schemas/PredictionLabel'
        confidence:
          type: number
          format: float
        real_label:
          $ref: '#/components/schemas/PredictionLabel'
          nullable: true
    PredictionUpdate:
      type: object
      properties:
        image_id:
          type: integer
        device_id:
          type: integer
        prediction_label:
          $ref: '#/components/schemas/PredictionLabel'
        confidence:
          type: number
          format: float
        real_label:
          $ref: '#/components/schemas/PredictionLabel'
          nullable: true
    FillLevel:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: integer
        last_update:
          type: string
          format: date-time
        fill_level_plastic:
          type: number
          format: float
          minimum: 0
          maximum: 1
        fill_level_paper:
          type: number
          format: float
          minimum: 0
          maximum: 1
        fill_level_waste:
          type: number
          format: float
          minimum: 0
          maximum: 1
      required: [device_id, last_update, fill_level_plastic, fill_level_paper, fill_level_waste]
    FillLevelUpdate:
      type: object
      properties:
        device_id:
          type: integer
        last_update:
          type: string
          format: date-time
        fill_level_plastic:
          type: number
          format: float
          minimum: 0
          maximum: 1
        fill_level_paper:
          type: number
          format: float
          minimum: 0
          maximum: 1
        fill_level_waste:
          type: number
          format: float
          minimum: 0
          maximum: 1
...
    User:
      type: object
      properties:
        id:
          type: integer
        user_name:
          type: string
        email:
          type: string
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        password:
          type: string
        creation_date:
          type: string
          format: date-time
          nullable: true
        admin:
          type: boolean
          nullable: true
          default: false
      required: [user_name, email]
    UserCreate:
      type: object
      required: [user_name, email, password]
      properties:
        user_name:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        creation_date:
          type: string
          format: date-time
          nullable: true
        admin:
          type: boolean
          nullable: true
          default: false
    UserUpdate:
      type: object
      properties:
        user_name:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        creation_date:
          type: string
          format: date-time
          nullable: true
        admin:
          type: boolean
          nullable: true
          default: false
...